Add PROXY protocol ALPN TLV helpers
